Migrate from J2E to JPA/Hibernate, Jetty & Jersey
Fix injections
Fix persistance transactions
Add custom EntityManagerFactories
Move packages out of y2018 subpackage
Remove unnecessary files
