+++ .gitignore
@@ -5,3 +5,4 @@
 .metadata/
 .settings/
 bin/
+*.log
