--- Iteration1.adoc
@@ -1,92 +1,0 @@
-= Itération 1 
-== Rappel travail à faire 
-
-Tâche 1 : ChessStateTxt 
-
--  Représenter et afficher un état du jeu ( = position des pions et joueur) grâce à une bibliotheque existante 
--  Choisir un format pour encoder ( ChessStateTxt ⇒ Format) et décoder ( Format ⇒ ChessStateTxt) un état de jeu  
-
-
-Tâche 2 : ChessMoveTxt 
-
-- Représenter un coup pour un joueur 
-- Calculer l’état résultant d’un coup joué sur un état 
-- Calculer le coup permettant de passer d’un état à un autre à partir de ces deux états 
-
-
-Tâche 3 : SetDB 
-
--  Construire un modèle relationnel pour stocker les états des parties 
--  Implémenter une entité JPA 
--  Implémenter les getters et setters 
-
-Tâche 4 : ShowClient 
-
--  Récupérer le dernier état d’un jeu à partir de son identifiant et l’afficher 
--  Récupérer un état precis d’un jeu à partir de l’identifiant de la partie et de l’identifiant de l’état 
-
-
-Description générale : Mise en place des bases du projet (état, mouvement et base de données) 
-
-
-== Realisation : 
-
-
-== Recherche en groupe de la librairie à utiliser.
-
-https://github.com/loopfz/JnegaChess 
-
-Dans un premier temps, nous avions choisi cette bibliothèque car elle nous semblait complète sans être trop complexe. Cependant, c’est un projet java et non un projet maven. Nous ne pouvons pas l’ajouter au projet en tant que bibliothèque. 
-
- 
-
-http://www.chesspresso.org/ 
-
-Cette bibliothèque avait particulièrement retenu notre attention. Elle semble très complète et riche. Elle est sous licence open-source et dispose d’une javadoc. Cette dernière est très bien réalisée (http://www.chesspresso.org/javadoc/index.html). JnegaChess a été préféré car plus simple.  
-
- 
-
-https://github.com/jvarsoke/ictk 
-
-Nous avons ensuite envisagé la librairie ictk. Celle-ci propose une représentation du jeu, des états et des mouvements ainsi que d’autres fonctionnalités qui pourraient nous être utiles. Cependant, il n’existe pas de release Maven pour cette librairie. De plus, les dernières mises à jour datent de 2004 et la librairie semble trop dense pour comprendre ses subtilités. 
-
-  
-
-https://github.com/marcelotomazini/chess 
-
-Cette bibliothèque ne donne pas d’informations sur la licence, donc on ne peut savoir si elle est compatible MIT ou non. De plus, du fait que ce soit un simple projet java et non Maven, il n’est pas également possible de l’ajouter en tant que bibliothèque. 
-
- 
-
-https://github.com/bhlangonijr/chesslib 
-
-Chesslib est une librairie régulièrement mise à jour, sous licence Apache. Elle propose une représentation complète du jeu d’échec. Cependant, l’affichage d’une partie est basé sur des charactères et non sur une interface graphique.  
-
-Chesslib nous semble être la meilleure solution pour notre projet, puisqu’il existe une dépendance Maven et qu’elle est facile à comprendre. Concernant la représentation graphique, celle-ci nous convient pour la première itération. Cependant, nous devrons sûrement créer une interface par la suite.  
-
-- Architecture générale 
-
-- "ChessStateTxt" et "ChessMoveTxt" réalisées en collaboration par Brassard Megan et Fremin du Sartel Thibaud 
-
-	   - Contributeur majeur de ChessStateTxt ( ChessState/ChessStateTest ) Thibaud Fremin du Sartel 
-	   - Contributeur majeur de ChessMoveTxt ( ChessMove/ChessMoveTest ) Megan Brassard. 
-
-
-
-- "SetDB" et "ShowClient" réalisées en collaboration par Lourenço Andréa et Douo Bougna Delmas 
-	  - Contributeur majeur de ShowClient (StateServlet ) Andréa Lourenço  
-	  - Contributeur majeur de SetDB (ChessStateEntity/ChessGameEntity) Delmas Douo Bougna 
-
-
-
-
-== Difficultés  rencontrées  : 
-
-	- Notion de "State" non présente dans la librairie
-	- Représentation des States dans la base de donnée : Nous avons réfléchi à plusieurs possibilités : +
-	A - Un état dépend d'une partie, la clé primaire dans la base de donnée est donc composée de l'identifiant de la partie et de l'identifiant de l'état +
-	B - Un état est indépendant d'une partie : On peut calculer tous les états possibles et les stocker avec un identifiant. On ajoute par la suite une table de jointure. +
-	C - Associer une liste d'identifiants State dans une partie. 
-	Nous avons opté pour la solution C. 
-	- Décision sur l'architecture : Nous souhaitions mettre en place une architecture qui soit assez facilement généralisable ( pour pouvoir implémenter un jeu de Poker plus tard par exemple) 
-	tout en étant simple. Nous nous sommes finalement focalisés sur la représentation du jeu d'echec. Nous sommes cependant conscients que le but de ce projet est de pouvoir généraliser le concept d'Assisted Board Game.
