+++ src/main/java/io/github/oliviercailloux/assisted_board_games/AppConfig.java
@@ -1,0 +1,15 @@
+package io.github.oliviercailloux.assisted_board_games;
+
+import org.glassfish.jersey.server.ResourceConfig;
+
+/**
+ * 
+ * @author theophile
+ *
+ */
+public class AppConfig extends ResourceConfig {
+
+    public AppConfig() {
+        packages("io.github.oliviercailloux.assisted_board_games.servlets");
+    }
+}
