+++ src/main/java/io/github/oliviercailloux/assisted_board_games/service/EntityManagerFactoryProducer.java
@@ -1,0 +1,27 @@
+package io.github.oliviercailloux.assisted_board_games.service;
+
+import java.util.HashMap;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Disposes;
+import javax.enterprise.inject.Produces;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
+
+/**
+ * From
+ * http://in.relation.to/2019/01/23/testing-cdi-beans-and-persistence-layer-under-java-se/
+ */
+@ApplicationScoped
+public class EntityManagerFactoryProducer {
+
+    @Produces
+    @ApplicationScoped
+    public EntityManagerFactory produceEntityManagerFactory() {
+        return Persistence.createEntityManagerFactory("ChessDB", new HashMap<>());
+    }
+
+    public void close(@Disposes EntityManagerFactory entityManagerFactory) {
+        entityManagerFactory.close();
+    }
+}
