+++ src/main/java/io/github/oliviercailloux/assisted_board_games/utils/JsonHelper.java
@@ -1,4 +1,4 @@
-package io.github.oliviercailloux.y2018.assisted_board_games.utils;
+package io.github.oliviercailloux.assisted_board_games.utils;
 
 import java.io.StringWriter;
