--- src/main/java/io/github/oliviercailloux/y2018/assisted_board_games/App.java
@@ -1,9 +1,0 @@
-package io.github.oliviercailloux.y2018.assisted_board_games;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("v1")
-public class App extends Application {
-
-}
